docs/wms que falhou/mais tentativa/backend/requirements.txt
fastapi[all]
sqlalchemy[asyncio]>=2.0.0
aiosqlite
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic-settings
pydantic[email]
uvicorn
